Added population matrix generator
Now, the load.py generates population.txt if it does not exist. The generate_population_matrix function also saves traj_trace.data which contains the coordinates of the atoms, and stacked_population.npy, which includes the population matrix at each time step. MSD and other quantities are not calculated as a part of population matrix calculations. So, no need to save duplicate info. There are some differences between the population matrices calculated from my code and the one we have in the database that can be related to wrap_atom_coords we use to wrap the coordinates we receive from LAMMPS. I will investigate it more and will let you know.