save population matrix files in local directory as well
I noticed received an error when testing a directory that did not have a population.txt. The reason is that we need a population matrix to calculate conductivity. Also, the htpmd should work fine on local directories alone. So, I added a couple of lines to save the population matrices in the local directory.

I noticed that this is not an efficient way to calculate the population matrix since it is very time-consuming. The best approach is to rewrite and modify the python2 code to have the same calculations in python3. I will take a look at it to see if I can do it without the pizza package.